% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_add_comments}
\alias{acro_add_comments}
\title{Add comments to outputs}
\usage{
acro_add_comments(name, comment)
}
\arguments{
\item{name}{The name of the output.}

\item{comment}{The comment.}
}
\description{
Add comments to outputs
}
\examples{
{
acro_init()
acro_crosstab(index=nursery_data$health, columns = nursery_data$finance)
acro_add_comments("output_0", "This is a crosstab")
}
}
