% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_custom_output}
\alias{acro_custom_output}
\title{Adds an unsupported output to the results dictionary}
\usage{
acro_custom_output(filename, comment = NULL)
}
\arguments{
\item{filename}{The name of the file that will be added to the list of the outputs.}

\item{comment}{An optional comment.}
}
\description{
Adds an unsupported output to the results dictionary
}
\examples{
{
acro_init()
acro_custom_output("XandY.jpeg")
}
}
