% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_disable_rounding}
\alias{acro_disable_rounding}
\title{Turns rounding off during a session}
\usage{
acro_disable_rounding()
}
\value{
No return value, called for side effects
}
\description{
Turns rounding off during a session
}
