% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_enable_rounding}
\alias{acro_enable_rounding}
\title{Turns rounding on during a session}
\usage{
acro_enable_rounding(base = NULL)
}
\arguments{
\item{base}{The base to round to}
}
\value{
No return value, called for side effects
}
\description{
Turns rounding on during a session
}
