% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_enable_suppression}
\alias{acro_enable_suppression}
\title{Turns suppression on during a session}
\usage{
acro_enable_suppression()
}
\value{
No return value, called for side effects
}
\description{
Turns suppression on during a session
}
