% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_finalise}
\alias{acro_finalise}
\title{Creates a results file for checking.}
\usage{
acro_finalise(path, ext)
}
\arguments{
\item{path}{Name of a folder to save outputs.}

\item{ext}{Extension of the results file. Valid extensions: {json, xlsx}.}
}
\description{
Creates a results file for checking.
}
\examples{
{
acro_init()
acro_finalise("results", "json")
}
}
