% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_regression.R
\name{acro_glm}
\alias{acro_glm}
\title{Fits Logit or Probit model.}
\usage{
acro_glm(formula, data, family)
}
\arguments{
\item{formula}{The formula specifying the model.}

\item{data}{The data for the model.}

\item{family}{Decide whether to fit a logit or probit model.}
}
\value{
Regression Results Wrapper
}
\description{
Fits Logit or Probit model.
}
\examples{
acro_init()
# extract relevant columns
df = charity_data[, c("survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs")]
# drop rows with missing values
df = df[complete.cases(df), ]
# convert survivor to numeric
df = transform(df, survivor = as.numeric(survivor))
# formula to fit
formula = "survivor ~ inc_activity + inc_grants + inc_donations + total_costs"
acro_glm(formula=formula, data=df, family="logit")
}
