% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_hist}
\alias{acro_hist}
\title{Histogram}
\usage{
acro_hist(
  data,
  column,
  breaks = 10,
  freq = TRUE,
  col = NULL,
  filename = "histogram.png"
)
}
\arguments{
\item{data}{The object holding the data.}

\item{column}{The column that will be used to plot the histogram.}

\item{breaks}{Number of histogram bins to be used.}

\item{freq}{If True, the result will contain the number of samples in each bin. If False, the result is the value of the probability density function at the bin.}

\item{col}{The color of the plot.}

\item{filename}{The name of the file where the plot will be saved.}
}
\value{
The histogram.
}
\description{
Histogram
}
