% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_init.R
\name{acro_init}
\alias{acro_init}
\title{Initialise an ACRO object}
\usage{
acro_init(
  config = "default",
  suppress = FALSE,
  mitigation = NULL,
  round_base = NULL,
  federated = NULL,
  envname = acro_venv,
  use_conda = NULL
)
}
\arguments{
\item{config}{Name of a yaml configuration file with safe parameters.}

\item{suppress}{Whether to automatically apply suppression.}

\item{mitigation}{The disclosure-control strategy applied to outputs, one of "none","suppress", "round".}

\item{round_base}{The base to round to when mitigation == "round".}

\item{federated}{Whether to run in federated mode.}

\item{envname}{Name of the Python environment to use.}

\item{use_conda}{Whether to use a Conda environment.
If \code{NULL}, looks for environment variable \code{ACRO_USE_CONDA},
defaults to \code{FALSE} if unset.}
}
\value{
Invisibly returns the ACRO object, which is used internally.
}
\description{
Initialise an ACRO object
}
