% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_regression.R
\name{acro_lm}
\alias{acro_lm}
\title{Fits Ordinary Least Squares Regression}
\usage{
acro_lm(formula, data)
}
\arguments{
\item{formula}{The formula specifying the model.}

\item{data}{The data for the model.}
}
\value{
Regression Results Wrapper.
}
\description{
Fits Ordinary Least Squares Regression
}
\examples{
acro_init()
# extract relevant columns
charity_data = charity_data[, c("inc_activity", "inc_grants")]
# drop rows with missing values
charity_data = charity_data[complete.cases(charity_data), ]
acro_lm(formula="inc_activity ~ inc_grants", data=charity_data)
}
