% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_pie}
\alias{acro_pie}
\title{Pie chart}
\usage{
acro_pie(
  data,
  column,
  radius = 0.8,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  col = NULL,
  border = NULL,
  lty = NULL,
  filename = "pie.png",
  ...
)
}
\arguments{
\item{data}{The object holding the data.}

\item{column}{The name of the column that will be used to plot the pie chart.}

\item{radius}{The radius of the pie chart.}

\item{clockwise}{logical indicating if slices are drawn clockwise or counter clockwise.}

\item{init.angle}{number specifying the starting angle (in degrees) for the slices. Defaults to 0 (i.e., ‘3 o'clock’) unless clockwise is true where init.angle defaults to 90 (degrees), (i.e., ‘12 o'clock’).}

\item{col}{colors to be used in filling or shading the slices}

\item{border}{The color to draw the border.}

\item{lty}{The line style.}

\item{filename}{The name of the file where the pie chart will be saved.}

\item{...}{Any other parameters.}
}
\value{
The pie chart
}
\description{
Pie chart
}
