% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_pivot_table}
\alias{acro_pivot_table}
\title{Pivot table}
\usage{
acro_pivot_table(
  data,
  values = NULL,
  index = NULL,
  columns = NULL,
  aggfunc = "mean"
)
}
\arguments{
\item{data}{The data to operate on.}

\item{values}{Column to aggregate, optional.}

\item{index}{If an array is passed, it must be the same length as the data. The list can contain any of the other types (except list). Keys to group by on the pivot table index. If an array is passed, it is being used as the same manner as column values.}

\item{columns}{If an array is passed, it must be the same length as the data. The list can contain any of the other types (except list). Keys to group by on the pivot table column. If an array is passed, it is being used as the same manner as column values.}

\item{aggfunc}{If list of strings passed, the resulting pivot table will have hierarchical columns whose top level are the function names}
}
\value{
Cross tabulation of the data.
}
\description{
Pivot table
}
\examples{
{
# Initilaise an ACRO object
acro_init()
# Call the acro_pivot_table function
table = acro_pivot_table(charity_data,
                         values= "inc_grants",
                         index="grant_type",
                         aggfunc="mean")
}
}
