% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_table}
\alias{acro_table}
\title{Compute a simple cross tabulation of two (or more) factors.}
\usage{
acro_table(index, columns, dnn = NULL, deparse.level = 0, ...)
}
\arguments{
\item{index}{Values to group by in the rows.}

\item{columns}{Values to group by in the columns.}

\item{dnn}{The names to be given to the dimensions in the result}

\item{deparse.level}{Controls how the default \code{dnn} is constructed.}

\item{...}{Any other parameters.}
}
\value{
Cross tabulation of the data
}
\description{
Compute a simple cross tabulation of two (or more) factors.
}
