% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{charity_data}
\alias{charity_data}
\title{A dataset containing sample data}
\format{
A data frame with columns:
\describe{
\item{}{}
\item{}{}

}
}
\source{
{}
}
\usage{
charity_data
}
\description{
This dataset includes sample data for demonstration purposes.
}
\examples{
data(charity_data)
}
\keyword{datasets}
