% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_factors}
\alias{create_factors}
\title{Convert data to factors and manage NA levels}
\usage{
create_factors(Values, useNA)
}
\arguments{
\item{Values}{to create factors on}

\item{useNA}{whether to include NA values in the table}
}
\value{
R vector factor
}
\description{
Convert data to factors and manage NA levels
}
