% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_init.R
\name{install_conda}
\alias{install_conda}
\title{Install ACRO in a conda environment}
\usage{
install_conda(envname)
}
\arguments{
\item{envname}{Name of the conda environment}
}
\description{
Install ACRO in a conda environment
}
