% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_init.R
\name{install_venv}
\alias{install_venv}
\title{Initialise ACRO in a Python virtual environment}
\usage{
install_venv(envname = acro_venv)
}
\arguments{
\item{envname}{Name of the virtual environment}
}
\description{
Initialise ACRO in a Python virtual environment
}
