% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_excluded}
\alias{is_excluded}
\title{Identify values to be excluded from an object}
\usage{
is_excluded(Values, exclude_list)
}
\arguments{
\item{Values}{to exclude from}

\item{exclude_list}{Items to exclude from the values}
}
\value{
logical vector that shows the places of the items that needs to be excluded
}
\description{
Identify values to be excluded from an object
}
