% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_invalid}
\alias{is_invalid}
\title{Identify empty values; NA or NaN in an object}
\usage{
is_invalid(Values)
}
\arguments{
\item{Values}{to check for empty values in it}
}
\value{
logical vector that shows the places of the empty values
}
\description{
Identify empty values; NA or NaN in an object
}
