% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_pandas_categorical}
\alias{to_pandas_categorical}
\title{Convert an R factor to a Pandas Categorical}
\usage{
to_pandas_categorical(Values, pd)
}
\arguments{
\item{Values}{R factor vector to be converted}

\item{pd}{Reference to the Python \code{pandas} module}
}
\value{
A Python \code{pandas.Categorical} object
}
\description{
Convert an R factor to a Pandas Categorical
}
