% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_add_exception}
\alias{acro_add_exception}
\title{Adds an exception request to an output.}
\usage{
acro_add_exception(name, reason)
}
\arguments{
\item{name}{The name of the output.}

\item{reason}{The comment.}
}
\description{
Adds an exception request to an output.
}
\examples{
{
acro_init()
acro_crosstab(index=charity_data[,c("year")], columns= charity_data[,c("grant_type")])
acro_add_exception("output_0", "This is not disclosive")
}
}
