% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_crosstab}
\alias{acro_crosstab}
\title{Compute a simple cross tabulation of two (or more) factors.}
\usage{
acro_crosstab(index, columns, values = NULL, aggfunc = NULL)
}
\arguments{
\item{index}{Values to group by in the rows.}

\item{columns}{Values to group by in the columns.}

\item{values}{Array of values to aggregate according to the factors. Requires \code{aggfunc} be specified.}

\item{aggfunc}{If specified, requires \code{values} be specified as well.}
}
\value{
Cross tabulation of the data
}
\description{
Compute a simple cross tabulation of two (or more) factors.
}
\examples{
{
# Initilaise an ACRO object
acro_init()
# Call the acro_crosstab function
table = acro_crosstab(index=nursery_data$health, columns = nursery_data$finance)
}
}
