% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_finalise}
\alias{acro_finalise}
\title{Creates a results file for checking.}
\usage{
acro_finalise(path, ext)
}
\arguments{
\item{path}{Name of a folder to save outputs.}

\item{ext}{Extension of the results file. Valid extensions are json or xlsx.}
}
\value{
No return value, called for side effects
}
\description{
Creates a results file for checking.
}
