% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_regression.R
\name{acro_glm}
\alias{acro_glm}
\title{Fits Logit or Probit model.}
\usage{
acro_glm(formula, data, family)
}
\arguments{
\item{formula}{The formula specifying the model.}

\item{data}{The data for the model.}

\item{family}{Decide whether to fit a logit or probit model.}
}
\value{
Regression Results Wrapper
}
\description{
Fits Logit or Probit model.
}
