% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_init.R
\name{acro_init}
\alias{acro_init}
\title{Initialise an ACRO object}
\usage{
acro_init(suppress = FALSE)
}
\arguments{
\item{suppress}{Whether to automatically apply suppression.}
}
\value{
No return value, called for side effects
}
\description{
Initialise an ACRO object
}
