% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_init.R
\name{acro_init}
\alias{acro_init}
\title{Initialise an ACRO object}
\usage{
acro_init(suppress = FALSE, envname = acro_venv, use_conda = FALSE)
}
\arguments{
\item{suppress}{Whether to automatically apply suppression.}

\item{envname}{Name of the Python environment to use.}

\item{use_conda}{Whether to use a Conda environment (\code{TRUE}) or venv (\code{FALSE}).}
}
\value{
Invisibly returns the ACRO object, which is used internally.
}
\description{
Initialise an ACRO object
}
