% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_print_outputs}
\alias{acro_print_outputs}
\title{Prints the current results dictionary.}
\usage{
acro_print_outputs()
}
\description{
Prints the current results dictionary.
}
\examples{
{
acro_init()
acro_crosstab(index=nursery_data$health, columns = nursery_data$finance)
acro_print_outputs()
}
}
