% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_remove_output}
\alias{acro_remove_output}
\title{Remove outputs}
\usage{
acro_remove_output(name)
}
\arguments{
\item{name}{Key specifying which output to remove, e.g., 'output_0'.}
}
\value{
No return value, called for side effects
}
\description{
Remove outputs
}
