% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_remove_output}
\alias{acro_remove_output}
\title{Remove outputs}
\usage{
acro_remove_output(name)
}
\arguments{
\item{name}{Key specifying which output to remove, e.g., 'output_0'.}
}
\description{
Remove outputs
}
\examples{
{
acro_init()
acro_crosstab(index=nursery_data$health, columns = nursery_data$finance)
acro_remove_output("output_0")
}
}
