% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_commands.R
\name{acro_rename_output}
\alias{acro_rename_output}
\title{Rename outputs}
\usage{
acro_rename_output(old, new)
}
\arguments{
\item{old}{The old name of the output.}

\item{new}{The new name of the output.}
}
\value{
No return value, called for side effects
}
\description{
Rename outputs
}
