% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_summarise}
\alias{acro_summarise}
\title{Title}
\usage{
acro_summarise(.data, ..., .groups = NULL, .by = NULL)
}
\arguments{
\item{.data}{A data frame or a data frame extension}

\item{...}{Name-value pairs of summary functions. The name will be the name of the variable in the result}

\item{.groups}{Grouping structure of the result}

\item{.by}{Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by()}
}
\value{
Summary of the data
}
\description{
Title
}
