% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acro_tables.R
\name{acro_summarise}
\alias{acro_summarise}
\title{Creates a new data frame. It returns one row for each combination of grouping variables; if there are no grouping variables, the output will have a single row summarising all observations in the input}
\usage{
acro_summarise(.data, ..., .groups = NULL, .by = NULL)
}
\arguments{
\item{.data}{A data frame or a data frame extension}

\item{...}{Name-value pairs of summary functions. The name will be the name of the variable in the result}

\item{.groups}{Grouping structure of the result}

\item{.by}{Optionally, a selection of columns to group by for just this operation, functioning as an alternative to group_by()}
}
\value{
Summary of the data
}
\description{
Creates a new data frame. It returns one row for each combination of grouping variables; if there are no grouping variables, the output will have a single row summarising all observations in the input
}
