% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{install_acro}
\alias{install_acro}
\title{Install the python library 'acro'}
\usage{
install_acro(..., envname = "r-acro")
}
\arguments{
\item{...}{Any other parameters.}

\item{envname}{the name of the python virtual environment}
}
\value{
No return value, called for side effects
}
\description{
Install the python library 'acro'
}
