% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtualenv.R
\name{install_acro}
\alias{install_acro}
\title{Install acro}
\usage{
install_acro(envname = "r-acro", ...)
}
\arguments{
\item{envname}{the name of the Python virtual environment}

\item{...}{Any other parameters.}
}
\value{
No return value, called for side effects
}
\description{
Install acro
}
