% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nursery_data.R
\docType{data}
\name{nursery_data}
\alias{nursery_data}
\title{Nursery Database}
\format{
A data frame with columns:
A data frame with 12960 rows and 9 columns:
\describe{
\item{parents}{Parents' occupation}
\item{has_nurs}{Child's nursery}
\item{form}{Form of the family}
\item{children}{Number of children}
\item{housing}{Housing conditions}
\item{finance}{Financial standing of the family}
\item{social}{Social conditions}
\item{health}{Health conditions}
\item{recommend}{The ranking of applications for nursery schools}

}
}
\source{
{https://www.openml.org/search?type=data&status=active&id=26&sort=runs}
}
\usage{
nursery_data
}
\description{
This dataset is originated from a hierarchical decision model created to evaluate applications for nursery schools.
}
\examples{
data(nursery_data)
}
\keyword{datasets}
